apply permission for api 2

+++ Absoft/Controllers/HangMucVatTuController.cs
@@ -1,7 +1,9 @@
-﻿using Absoft.Extentions;
+﻿using Absoft.Authorization;
+using Absoft.Extentions;
 using Absoft.Helpers;
 using Absoft.Repositories.Interfaces;
 using Absoft.ViewModels;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
 
@@ -10,14 +12,22 @@ namespace Absoft.Controllers
     public class HangMucVatTuController : BaseController
     {
         IHangMucVatTuRepository _hangMucVatTuRepository;
-        public HangMucVatTuController(IHangMucVatTuRepository hangMucVatTuRepository)
+        private readonly IAuthorizationService _authorizationService;
+
+        public HangMucVatTuController(IHangMucVatTuRepository hangMucVatTuRepository,
+            IAuthorizationService authorizationService)
         {
             _hangMucVatTuRepository = hangMucVatTuRepository;
+            _authorizationService = authorizationService;
         }       
 
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            //var auth = await _authorizationService.AuthorizeAsync(User, "HANG_MUC_VAT_TU", Operations.Read);
+            //if (auth.Succeeded == false)
+            //    return Unauthorized();
+
             var models = await _hangMucVatTuRepository.GetAllAsync();
             return Ok(models);
         }
@@ -25,6 +35,10 @@ namespace Absoft.Controllers
         [HttpGet("getAllPaging")]
         public async Task<IActionResult> GetAllPaging([FromQuery]PagingParams pagingParams)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "HANG_MUC_VAT_TU", Operations.Read);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var paged = await _hangMucVatTuRepository.GetAllPagingAsync(pagingParams);
             Response.AddPagination(paged.CurrentPage, paged.PageSize, paged.TotalCount, paged.TotalPages);
             return Ok(paged.Items);
@@ -33,6 +47,10 @@ namespace Absoft.Controllers
         [HttpGet("{id}")]
         public async Task<IActionResult> GetDetail(int id)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "HANG_MUC_VAT_TU", Operations.Read);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var models = await _hangMucVatTuRepository.GetByIdAsync(id);
             return Ok(models);
         }
@@ -40,12 +58,20 @@ namespace Absoft.Controllers
         [HttpPut]
         public async Task<IActionResult> Update(HangMucVatTuViewModel hangMucVatTuViewModel)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "HANG_MUC_VAT_TU", Operations.Update);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var result = await _hangMucVatTuRepository.UpdateAsync(hangMucVatTuViewModel);
             return Ok(result);
         }
         [HttpPost]
         public async Task<IActionResult> Insert(HangMucVatTuViewModel hangMucVatTuViewModel)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "HANG_MUC_VAT_TU", Operations.Create);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var id = await _hangMucVatTuRepository.CheckTonTai(hangMucVatTuViewModel.TenHM);
             if (id == -1)
             {
@@ -60,6 +86,10 @@ namespace Absoft.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var auth = await _authorizationService.AuthorizeAsync(User, "HANG_MUC_VAT_TU", Operations.Delete);
+            if (auth.Succeeded == false)
+                return Unauthorized();
+
             var result = await _hangMucVatTuRepository.DeleteAsync(id);
             return Ok(result);
         }
